requirements.txt
amqp==2.5.0
astroid==2.2.5
backcall==0.1.0
beautifulsoup4==4.7.1
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.1
django-allauth==0.39.1
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-cors-headers==3.0.1
django-debug-toolbar==2.0
django-extensions==2.2.1
django-filter==2.1.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-timezone-field==3.0
djangorestframework==3.9.4
drf-access-policy==0.4.2
idna==2.8
ipdb==0.12.2
ipython==7.7.0
ipython-genutils==0.2.0
isort==4.3.20
jedi==0.15.1
kombu==4.6.3
lazy-object-proxy==1.4.1
lxml==4.3.0
Markdown==3.1.1
mccabe==0.6.1
oauthlib==3.0.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.3.1
python-crontab==2.3.8
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
traitlets==4.3.2
typed-ast==1.3.5
urllib3==1.24.3
vine==1.3.0
wcwidth==0.1.7
wrapt==1.11.1
